Add menu option to display stream information
During playback listen for changes in stream properties
(bitrate, codec, etc.) that are sent from MPV and display
them in a new stream information window.

Stream info is only available when a stream is being played
and is accessible using the F2 shortcut key. The bitrate
will be averaged over the last 30 seconds and the UI will
be updated every 5 seconds.