inhibit: Remove configuration
The app tells the session manager it's doing something, it's up to the
session manager to offer configuration for this if need be or to provide
ways of handling event to the user. The app itself should always do
this.

Note that we don't have any such settings in rhythmbox, celluloid etc..